NOISSUE: Move SyncT to testutuils
