+++ logicrunner/shutdown/flag_test.go
@@ -20,15 +20,16 @@ import (
 	"testing"
 	"time"
 
+	"github.com/insolar/insolar/testutils"
+
 	"github.com/stretchr/testify/require"
 
-	"github.com/insolar/insolar/insolar/utils"
 	"github.com/insolar/insolar/instrumentation/inslogger"
 )
 
 func TestFlag(t *testing.T) {
 	ctx := inslogger.TestContext(t)
-	syncT := utils.SyncT{T: t}
+	syncT := testutils.SyncT{T: t}
 
 	flag := NewFlag()
